changed folder structure to clean
added some test code for future but doesnt really work. will work on it
later
