+++ app/src/main/java/com/example/tic_tac_toechallenge/presentation/MainScreen.kt
@@ -1,4 +1,4 @@
-package com.example.tic_tac_toechallenge
+package com.example.tic_tac_toechallenge.presentation
 
 
 import android.util.Log
@@ -10,13 +10,13 @@ import androidx.compose.material3.*
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.example.tic_tac_toechallenge.R
+import com.example.tic_tac_toechallenge.presentation.sign_in.UserData
 import com.example.tic_tac_toechallenge.ui.theme.TicTacToeChallengeTheme
 
 @Composable
