+++ app/src/main/java/com/example/tic_tac_toechallenge/presentation/sign_in/SignInResult.kt
@@ -1,11 +1,10 @@
-package com.example.tic_tac_toechallenge
+package com.example.tic_tac_toechallenge.presentation.sign_in
 
 data class SignInResult(
     val data: UserData?,
     val errorMessage: String?,
 
-)
-
+    )
 
 data class UserData(
     val userId: String,
