+++ app/src/main/java/com/example/tic_tac_toechallenge/presentation/sign_in/SignInState.kt
@@ -1,4 +1,4 @@
-package com.example.tic_tac_toechallenge
+package com.example.tic_tac_toechallenge.presentation.sign_in
 
 import java.lang.Error
