+++ app/src/main/java/com/example/tic_tac_toechallenge/presentation/sign_in/SignInViewModel.kt
@@ -1,6 +1,8 @@
-package com.example.tic_tac_toechallenge
+package com.example.tic_tac_toechallenge.presentation.sign_in
 
 import androidx.lifecycle.ViewModel
+import com.example.tic_tac_toechallenge.presentation.sign_in.SignInResult
+import com.example.tic_tac_toechallenge.presentation.sign_in.SignInState
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
@@ -8,7 +10,6 @@ import kotlinx.coroutines.flow.update
 class SignInViewModel: ViewModel() {
     private val _state = MutableStateFlow(SignInState())
     val state  = _state.asStateFlow()
-
 
     fun onSignInResult(result: SignInResult) {
         _state.update { it.copy(
